Wall moves back and forth now.
I can make the wall be higher up, but trying to make the wall be closer
to the ground has no effect. decreasing the width of the wall lets me
put it closer down. I changed the height of the wall to be the same as
that width, and sure enough, it was touching, but not actually below the
floor. Making the height of the wall to be larger than the width, allows
you to put the wall partially into the floor.  Apparently ACframework is
using width to check to make sure that the wall is above ground, instead
of height.