Tweaked critter speeds to add character.
Snakes are more like guards guarding key points, such as keys and doors.
BigHead relentlessly chases the player no matter where the player is.
Sailormoon, the chicken, and robot, are like the medium enemies chasing
the playerwise, though the robot may be changed to focus more on
shooting than chasing the player. Damage will need to be tweaked
accordingly.