Fixed egg and poisonsplat sounds not working
The player now actually gets poisoned by the poison bullets. The fix for
this was the same as the fix for the sound effect not working. The
bullet collided with the player, but deleted itself before the player
could do it's collision with the bullet.

Checking if the critter is dead moved to the top of the collide function
and one small error fixed.
Removed the sound when the player jumps on a critter.